Fix duplicate entries being created in users
If a user had not previously logged into the app via Facebook, there
would be an entry created in the users array (which is used for
looking up other users). However, there was an issue whereby even if
a user had previously logged into the app via Facebook and their
details had been stored, there was another entry created into the
users array each time the user signed in with Facebook.